requirements.txt
alembic==1.7.4
anyio==3.3.4
apache-airflow==2.2.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
blinker==1.4
cattrs==1.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cryptography==35.0.0
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.1.0
Flask==1.1.4
Flask-AppBuilder==3.3.4
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.17
gunicorn==20.1.0
h11==0.12.0
htmlmin==0.1.12
httpcore==0.13.7
httpx==0.20.0
idna==3.3
ImageHash==4.2.1
inflection==0.5.1
ipykernel==6.5.0
ipython==7.29.0
iso8601==0.1.16
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.6
jupyter-core==4.9.1
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
lockfile==0.12.2
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
missingno==0.5.0
multimethod==1.6
nest-asyncio==1.5.1
networkx==2.6.3
numpy==1.21.3
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
openpyxl==3.0.9
packaging==21.0
pandas==1.3.4
pandas-profiling==3.1.0
parso==0.8.2
pendulum==2.1.2
petl==1.7.4
phik==0.12.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.3.1
prison==0.2.1
prompt-toolkit==3.0.22
psutil==5.8.0
psycopg2==2.9.1
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyWavelets==1.1.1
pywin32==302
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
rfc3986==1.5.0
rich==10.12.0
scikit-learn==1.0.1
scikit-surprise==1.1.1
scipy==1.7.1
seaborn==0.11.2
setproctitle==1.2.2
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.9
surprise==0.1
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.10.0.2
uber-rides==0.6.0
unicodecsv==0.14.1
urllib3==1.26.7
visions==0.7.4
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
